+++ src/counterfactuals/info_extraction.jl
@@ -102,3 +102,5 @@ function target_encoded(ce::CounterfactualExplanation, data::CounterfactualData)
 end
 
 target_encoded(ce::CounterfactualExplanation) = target_encoded(ce, ce.data)
+
+num_counterfactuals(ce::CounterfactualExplanation) = ce.num_counterfactuals