+++ src/evaluation/Evaluation.jl
@@ -32,6 +32,7 @@ export faithfulness
 export plausibility_measures, default_measures, distance_measures, all_measures
 export concatenate_benchmarks
 export compute_divergence
+export feature_sensitivity
 
 "Available plausibility measures."
 const plausibility_measures = [